moved test configuration and dependency files to js_tests folder
