+++ tests/js_tests/.babelrc
