+++ tests/js_tests/jest.config.js
@@ -1,13 +1,13 @@
 module.exports = {
     testEnvironment: "", 
-    roots: ["<rootDir>/js-tests"],
-    setupFilesAfterEnv: ["<rootDir>/js-tests/setup.js"],
+    roots: ["<rootDir>"],
+    setupFilesAfterEnv: ["<rootDir>/setup.js"],
     collectCoverage: true,    
     coverageDirectory: "jest_coverage", 
     coverageReporters: ["text", "lcov"], 
     moduleFileExtensions: ["js", "jsx", "json", "node"],
     moduleNameMapper: {
-      "^cesium$": "<rootDir>/js-tests/__mocks__/cesium.js",
+      "^cesium$": "<rootDir>/__mocks__/cesium.js",
     },
     transformIgnorePatterns: ["/node_modules/"],
   };
